+++ rollup.config.js
@@ -1,6 +1,5 @@
 import resolve from 'rollup-plugin-node-resolve';
 import babel from 'rollup-plugin-babel';
-import replace from 'rollup-plugin-replace';
 import { terser } from 'rollup-plugin-terser';
 
 const env = process.env.NODE_ENV;
@@ -11,33 +10,28 @@ const config = {
       exports: 'named',
       file: 'lib/storage.umd.js',
       format: 'umd',
+      sourcemap: true,
       name: 'storage',
     },
     {
       exports: 'named',
       file: 'lib/storage.cjs.js',
+      sourcemap: true,
       format: 'cjs',
     },
     {
       file: 'lib/storage.esm.js',
+      sourcemap: true,
       format: 'esm',
-    }
+    },
   ],
   plugins: [
     resolve(),
     babel({
       exclude: 'node_modules/**',
-      plugins: ['@babel/external-helpers'],
-      externalHelpers: true
     }),
-    replace({
-      'process.env.NODE_ENV': JSON.stringify(env)
-    })
+    terser(),
   ],
 };
-
-if (env === 'production') {
-  config.plugins.push(terser());
-}
 
 export default config;